Add new states and attributes for the poly and object aswell

+++ src/graphics/Rasterizer.cpp
@@ -127,8 +127,8 @@ inline void Rasterizer::draw_scanline(const Edge &left, const Edge &right, int j
         if (p_z_buffer[index] > depth) {
             float z = 1.0f / one_over_z;
 
-            int src_x = (int)((text_coord_x * z) * (texture.width - 1) + 0.5f);
-            int src_y = (int)((text_coord_y * z) * (texture.height - 1) + 0.5f);
+            int src_x = (text_coord_x * z) * (texture.width - 1) + 0.5f;
+            int src_y = (text_coord_y * z) * (texture.height - 1) + 0.5f;
 
             uint32_t value = texture.get_pixel(src_x, src_y, light_amount).value;
